sw: give the 'Default Style' char style a programmatic name
So that referring to that style is possible from UNO API client code in
a way that is portable across multiple locales.

This also improves consistency between the default para and default char
styles, as the default para style already had this behavior.

Also add a new CppunitTest_sw_unowriter for such tests which assert the
UNO API from cppunit, instead of hacking this into
CppunitTest_sw_uiwriter.

Change-Id: I0a2b02378dce53c6b79c57780d1b7f14e89242e8
Reviewed-on: https://gerrit.libreoffice.org/57191
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins