Resolves: tdf#115950 proper double click return and bail out, tdf#117063
    commit 821994b0e0a941654fc77ff9087f337349b173b8
    CommitDate: Sun Apr 29 08:46:46 2018 +0200

        tdf#117063 Modify tree list double click behavior in the navigator

changed the return value in SvTreeListBox::DoubleClickHdl() from

  return !aDoubleClickHdl.IsSet() || aDoubleClickHdl.Call(this);

to an unconditional true. Earlier there was

    commit e7566ac0c0ec32eb8bee443e7ef52226f25194e3
    CommitDate: Tue Mar 13 16:28:40 2018 +0100

        tdf#116334: Actually when there is no handler, we have to return 'true'.

    -    aDoubleClickHdl.Call( this );
    -    return false;
    +    return !aDoubleClickHdl.IsSet() || aDoubleClickHdl.Call(this);

and before that

    commit db9f5c18302790f1cec0b150949adee6b317f270
    CommitDate: Thu Mar 1 15:41:13 2018 +0100

        tdf#115950: Indicate that the dialog was already destroyed.

    -    return true;
    +    return false;

Neither a constant false or true are correct return values here,
but only the value the double click handler, if any, returned to
indicate whether processing should continue (true) or not (false).

If handlers don't return a proper value so the intended behaviour
for tdf#117063 or anything else does not work then fix the
handlers instead.

If the handler returned false then don't even attempt to access
anything in SvImpLBox::MouseButtonDown() because an OK handler may
have destroyed everything and all is rotten.

Change-Id: Ia90c21288bedd7e5078dbe4b3dd6d9f5199a2a98
Reviewed-on: https://gerrit.libreoffice.org/57225
Tested-by: Jenkins
Reviewed-by: Jan Holesovsky <kendy@collabora.com>
Reviewed-by: Eike Rathke <erack@redhat.com>