tdf#72560 writerfilter: StyleProperty - also check DocDefaults
Styles without parents still have inheritance - from the DocDefaults.
So, make sure to check those properties too in
GetPropertyFromStyleSheet() (and therefore also in GetAnyProperty).

Currently, this should only impact when InStyleSheetImport(),
but if the docDefaults ever find a way to naturally migrate
into LO as doc defaults and not into the style itself, then
this will also be useful to check for paragraph properties etc.

Change-Id: Ic02cd10c104c0330684308f1380a77009206a664
Reviewed-on: https://gerrit.libreoffice.org/57805
Reviewed-by: Justin Luth <justin_luth@sil.org>
Tested-by: Justin Luth <justin_luth@sil.org>