Related: tdf#118735 WEBSERVICE preserve the hybrid string result, if possible
Possible only for constant literal URI arguments, as otherwise we
wouldn't know whether the hybrid string result actually matches
the call, and possible only for the very first recalc as the
hybrid string result will be replaced by the calculated result.

However, executing an initial recalc is necessary to add the (then
determined and maybe calculated) URI to the link manager after
having loaded a document.

Change-Id: I67738c13ecce54b43ebb52ea8a8282aebee5f03a
Reviewed-on: https://gerrit.libreoffice.org/57696
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Jenkins