tdf#118107 canvas opengl: avoid assertion failure with negative widths
In case an Impress shape has a "Fade in and Swivel" animation attached,
canvas calls OutputDevice::DrawBitmapEx() with a negative width. This
results in a call to OutputDevice::DrawDeviceAlphaBitmap(), which
asserts that the height/width is not negative.

Fix the problem by transforming the bitmap before calling
OutputDevice::DrawBitmapEx() in the GL case, similarly to how "complex"
transformations are handled already.

Change-Id: I65ccc8a984132c5921d6096bfe9c7a8fcfacd8dd
Reviewed-on: https://gerrit.libreoffice.org/57774
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins