tdf#118966 vcl: add a flag to determine if AA of fonts is used from the system
This is on by default (as there are loads of vcl users who expect font
AA working even if the default is off and they don't enable it), but the
svx UnoGraphicExporter disables it to make everyone happy.

The reason in practice AA was on by default is that the gtk backend uses
the system settings in GtkInstance::GetCairoFontOptions() (and not the
AA setting from the UI), and lclGetSystemTextAntiAliasMode() does the
same on Windows (at least in the direct write case). So now these
defaults again have higher priority than leaving the vcl-level default
unchanged.

Change-Id: I81267c0b036211525ac02d3282fa89d75510f4a8
Reviewed-on: https://gerrit.libreoffice.org/58199
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins