[API CHANGE] add some more asserts to the string functions
rtl_[u]String_alloc now requires that the length be >= 0.
Since this function is only @since Libreoffice 4.1, it is unlikely
to be widely used externally.

Removed some unit tests that were testing invalid or out of
range paramers, which are already not allowed according to the
documented contract of those functions.

The change in writerfilter is because the new asserts triggered
when running testFdo74745

The change in SwTextNode::EraseText is because testFdo60842
triggered the assert in replaceAt.

The change in SwFieldSlot::SwFieldSlot is because
testMoveRange::Import_Export_Import triggered the assert in
replaceAt.

The changes in SwFieldSlot::SwFieldSlot and TabControl::ImplGetItemSize
are due to failures in the uitests.

Change-Id: Ib317261067649b0de96df12873ce31360cd24681
Reviewed-on: https://gerrit.libreoffice.org/58390
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Noel Grandin <noel.grandin@collabora.co.uk>