Translate German variable names
This is not a literal translation. Rather, it follows the notions used in the UI,
as introduced in commit 35c24fc14054b5f7f2334e8f931a9f143685f51a - see
https://gerrit.libreoffice.org/gitweb?p=core.git;a=blob;f=cui/source/tabpages/labdlg.cxx;h=41bc3d72ea50e94d31aab369ea5dcdedbb00ddac;hb=35c24fc14054b5f7f2334e8f931a9f143685f51a#l86 .

This seem to be the legacy callouts described in https://help.libreoffice.org/Common/Callout .

Further similar translations in the same files will be done separately later.

Change-Id: Iac5f1e34f28f12cd24001c488d35ae3ba3996aea
Reviewed-on: https://gerrit.libreoffice.org/58584
Tested-by: Jenkins
Reviewed-by: Michael Stahl <Michael.Stahl@cib.de>