tdf#117994 ww8export: extend Chp over CR
Plan to revert this within a month - just a test and see.

MS Word has a placeholder that can receive chacter formatting
for the carriage return, but LO doesn't really. So normally
when LO exports, the CR is "unformated", which really
can affect things like paragraph size, or numbering.
So, extend the current character formatting to cover
the CR, so that what we see in Writer is what we get in Word.

This will cause regressions as Writer round-trips documents
that it incorrectly reads. (In other words, even though they
don't look right in Writer, they still used toround-tripped
OK, but now they will be saved with LO's distortions.)
Since normally if we *save* in writer, that implies that
changes have been made, and theoretically the document
looks the way the user wants to re-preserve it, it
is better to now have the document look as Writer
sees it than to preserve some prior formatting.

This really breaks unit test tdf102334.doc, but that
round-tripped document is already *really* broken in Word.
With this patch, it looks like the original again.
(tdf102334.doc looks bad in LO no matter what, it just looks
worse in this case.)

Change-Id: I04d3fdc7dbc66f8732456d7ac5b309c454992d8b
Reviewed-on: https://gerrit.libreoffice.org/58594
Tested-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Justin Luth <justin_luth@sil.org>