tdf#119169 impress, crash when editing bulleted list
regression from
    commit 0463c65b58e99cd31e4ea6f540c0f76e1bb89f21
    tdf#108608 Draw file unresponsive on large text pasted into textbox

Somewhere up in the accessibility code it has a stale paragraph index.
Judging up other code in this class, that is fairly common, so just add
similar safety checks.
The old code (pre-regression) only worked because it iterated over all
paragraphs, and it just ignored the paragraph index.

Change-Id: I51c573b95f976fa87228e4f7e9f33bd6ea0dc3c8
Reviewed-on: https://gerrit.libreoffice.org/58836
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>