tdf#101242 Support draw:display and draw:protect attributes of ODF
LibreOffice writes the properties visible, printable and locked of
layers into items in the subfile settings.xml and handles them as
properties of the view. ODF handles them as property of the layer.
To become more ODF conform as a first step these properties are now
read and written. They are used to initialize the view in case they
are missing in settings.xml, which is the case for foreign documents.
The ODF properties are written in addition to the items in
settings.xml, so that older versions will notice no difference with
such documents. SdModelTestBase is changed to handle odg as Draw
document.

Change-Id: I190ecc51fc6ee91ec4b96d06bb216ce517bdfcfe
Reviewed-on: https://gerrit.libreoffice.org/59269
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>