tdf#116370 cleanup Writer idle job handing
This prevents the start of the idle job, while processing itself,
so the fixed WinSalInstance::AnyInput of commit c2b19b6aceb4
("tdf#112975 WIN correctly handle VclInputFlags::OTHER") won't
report the timer events of the re-started idle job to process.

Fixes the early abort of the background job, which resulted in
the busy loop of the reported bug and this strange printing
behaviour.

P.S. I'm not sure, why this was just broken on Windows.

Change-Id: I6503dcd925c9a0ed843e794a31eea32a4a4b2889
Reviewed-on: https://gerrit.libreoffice.org/59279
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>