tdf#119282 Area tab, Pattern color change
regression from

    commit 08901f740652748916d2a78c020d4198e376d126
	use VirtualDevice in createHistorical8x8FromArray

Move the code inside vcl so we don't need to expose BitmapWriteAccess.

I tried to detect this thing by counting the number of independent
colors in the bitmap, but that didn't work. It is used from more than
one place, and appears to be very determined that it needs a 2-color
palette to work properly.

Change-Id: Id11dd9ea78e5e522a6083d6a799e801cac81fd5b
Reviewed-on: https://gerrit.libreoffice.org/59331
Reviewed-by: Michael Meeks <michael.meeks@collabora.com>
Tested-by: Jenkins