tdf#82177 ooxmlimport: ignore direct insideV/H cell borders
17.4.25 insideV (Table Cell Inside Vertical Edges Border)
This element specifies the border which shall be displayed on
all interior vertical edges of the current group of table cells.
[Note: Although individual table cells have no concept
of n internal edge, which would render this property useless
in most cases, it is used to determine the cell borders to
apply to a specific group of cells as part of table
conditional formatting in a table style, for example, the
inside vertical edges on the set of cells in the header row.
end note]

So, I interpret this as insideV/H having meaning only within table
styles, and not when directly applied to a cell. The only
documents I've seen with insideV/H directly applied to a cell seem to
have been created by LO - which dumps them everywhere, redundantly.

Change-Id: Ie76e9af6845cc87b0b847050b031e57d95b1a31b
Reviewed-on: https://gerrit.libreoffice.org/59674
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>