tdf#104354 writerfilter: rewrite Autospacing
which also resolves a regression from tdf#113258.

The previous code segments were being spread out all over
the place. It all consolidates nicely in finishParagraph,
and the code is much easier to read using the new
GetAnyProperty function.

Plus there were regressions creeping in.
The mere presense of the Autospacing property normally, but
not necessarily means that it is on. Verify that it is enabled,
and update grabbag if autoSpacing changes.

Additionally, support was added for a zero top margin for
the first paragraph of the document.

This will be too hard to backport since it depends on
2278c2fb50b4f68772e27cc6cdf1f060a7fd8292 from tdf#72560
and the many code fixes related to GetPropertyFromStyleSheets.

Change-Id: Iaf1600fffea54e9800e215e89cad40006d5bcdda
Reviewed-on: https://gerrit.libreoffice.org/59705
Reviewed-by: László Németh <nemeth@numbertext.org>
Tested-by: László Németh <nemeth@numbertext.org>