tdf#104354 writerfilter: not FirstParaInSection if InFootOrEndnote
or InComment. The first footnote does not get special treatment in
autospacing either, so no exception needs to be made for it like in
autospacing, unlike many other "firsts".

The unit test is simply preventative. I'm a bit surprised,
but the section isn't even marked as "IsFirstSection" until after
the comment and footnote paragraphs are finished. That seems a bit
odd to me, since they are in the FirstSection after all.

Change-Id: Ib0107c969bf624e44f0c78b936bf9531fd25bfc2
Reviewed-on: https://gerrit.libreoffice.org/59858
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: László Németh <nemeth@numbertext.org>