tdf#94009 ww8export: always export section margins
Since the internal SW defaults (0) don't match the import
defaults, always write the top/bottom, left/right margins
into the document definition.

It is very rare indeed to have a zero margin anyway, since
the page margin being zero is highly discouraged because
of printing. The bug report is for DOCX, but it also
affects DOC.

I don't have an example where LRSpace is skipped, but it
only makes sense to treat these two the same just in case.

Change-Id: Ie9a08ad0dd4f73bc976756fe244fc33e2dc804f3
Reviewed-on: https://gerrit.libreoffice.org/59967
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>