tdf#119340 WIN don't always re-init the HFONT
After all the LogicalFontInstance refactoring, its FontSelectData
is now immutable, so the HFONT will never need to be changed.
Therefore we can skip the HFONT lookup in WinSalGraphics::SetFont,
if the WinFontInstance already has it set.

All the HFONT handling (basically WinSalGraphics::ImplDoSetFont)
should be moved to the WinFontInstance to further clean this up.

Change-Id: I1c9a9e3414677d77f0f007d7ace653bb53b8474d
Reviewed-on: https://gerrit.libreoffice.org/59988
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>