tdf#117721 draw ui: add .uno::SetOptimalColumnWidth
Optimize column width: Adjusts the width of the
selected columns to fit the content,
without changing the width of the table.
Any leftover space is distributed proportionately,
with thin columns growing slightly,
and wide columns growing much wider.

The implementation and results are different from how Writer
does it, but tables themselves are also very different in Draw.
So I don't think it needs to be handled identically.

Writer's implementation is really "minimalColumnWidth" anyway.

Change-Id: Ia10cfa9822d7eef3c4909a82c21535aa6668b143
Reviewed-on: https://gerrit.libreoffice.org/60078
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>