sd: support inserting background image from file via .uno:SelectBackground
The UNO command always prompted the user via file open dialog.
This change allows for passing the filename as an argument
to allow for inserting slide background programatically.

Also, hide the Insert Image button in the sidebar
since we can't use that (just yet), because
it invokes .uno:SelectBackground directly. We would
need to send LOOL a notification to invoke this on
its own end to first prompt the user for a file.

(cherry picked from commit 2118143bdd246921439ba9e835207585203dd45f)

Change-Id: I20c0e33d66f8bcd72a6388e39c4ac92e64978f45
Reviewed-on: https://gerrit.libreoffice.org/73481
Tested-by: Jenkins
Reviewed-by: Jan Holesovsky <kendy@collabora.com>