tdf#115094 part I: do not move graphic nodes
Do not move graphic nodes in SwXText::convertToTextFrame()
since it is not reasonable (they are anchored to something that
already moved) and hence it causes errors.

A static function checks wether the current node is a graphic.

A small change in writerfilter was done since a unit test failed:
The condition "m_nAnchorType != 0"  was removed since it is not
reasonable at this place.

Change-Id: I8f27985f6f6c8329483370a7b8baaf9949513f1b
Reviewed-on: https://gerrit.libreoffice.org/60860
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>