tdf#64242 sw optimal row height, not minimize
The previous optimal implementation was cloned as a new
uno:MinimalColumnWidth function.

Optimal should try to keep the table the same size, but ultimately
the ideal row/table height will be balanced, so grow to achieve
that if necessary.

Optimize row height: Adjust the height of the selected rows
to match the height of the tallest row in the selection
(fit to content), without shrinking the table.
This option is the same as minimizing row height and then
distributing rows evenly except that it adds the benefit
of preventing the table from shrinking.

Change-Id: Ib772ae4943127d17a915cdae8cd6000db3a9d164
Reviewed-on: https://gerrit.libreoffice.org/60903
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>